requirements.txt
python-telegram-bot==21.0.1
openai==1.3.5
python-dotenv==1.0.0 